+++ CHANGES.rst
@@ -38,6 +38,7 @@ development
 - Protect sensible configuration settings from leaking into javascript environment
 - Use vendor information from application settings
 - Refactor development mode flag
+- Release packaging wrt webpack
 
 2017-09-08 0.135.0
 ==================